daemon: move postSnap and inst.dispatch tests to api_snaps_test.go
