+++ daemon/api_test.go
@@ -26,7 +26,6 @@ import (
 	"fmt"
 	"mime/multipart"
 	"net/http"
-	"net/http/httptest"
 	"strings"
 
 	"gopkg.in/check.v1"
@@ -36,7 +35,6 @@ import (
 	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
-	"github.com/snapcore/snapd/sandbox"
 	"github.com/snapcore/snapd/snap"
 	"github.com/snapcore/snapd/snap/channel"
 	"github.com/snapcore/snapd/snap/snaptest"
@@ -448,717 +446,6 @@ func (s *apiSuite) TestUserFromRequestHeaderValidUser(c *check.C) {
 	c.Check(user, check.DeepEquals, expectedUser)
 }
 
-func (s *apiSuite) TestPostSnapBadRequest(c *check.C) {
-	buf := bytes.NewBufferString(`hello`)
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result, check.NotNil)
-}
-
-func (s *apiSuite) TestPostSnapBadAction(c *check.C) {
-	buf := bytes.NewBufferString(`{"action": "potato"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result, check.NotNil)
-}
-
-func (s *apiSuite) TestPostSnapBadChannel(c *check.C) {
-	buf := bytes.NewBufferString(`{"channel": "1/2/3/4"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result, check.NotNil)
-}
-
-func (s *apiSuite) TestPostSnap(c *check.C) {
-	s.testPostSnap(c, false)
-}
-
-func (s *apiSuite) TestPostSnapWithChannel(c *check.C) {
-	s.testPostSnap(c, true)
-}
-
-func (s *apiSuite) testPostSnap(c *check.C, withChannel bool) {
-	d := s.daemonWithOverlordMock(c)
-
-	soon := 0
-	ensureStateSoon = func(st *state.State) {
-		soon++
-		ensureStateSoonImpl(st)
-	}
-
-	s.vars = map[string]string{"name": "foo"}
-
-	snapInstructionDispTable["install"] = func(inst *snapInstruction, _ *state.State) (string, []*state.TaskSet, error) {
-		if withChannel {
-			// channel in -> channel out
-			c.Check(inst.Channel, check.Equals, "xyzzy")
-		} else {
-			// no channel in -> no channel out
-			c.Check(inst.Channel, check.Equals, "")
-		}
-		return "foooo", nil, nil
-	}
-	defer func() {
-		snapInstructionDispTable["install"] = snapInstall
-	}()
-
-	var buf *bytes.Buffer
-	if withChannel {
-		buf = bytes.NewBufferString(`{"action": "install", "channel": "xyzzy"}`)
-	} else {
-		buf = bytes.NewBufferString(`{"action": "install"}`)
-	}
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-	c.Check(chg.Summary(), check.Equals, "foooo")
-	var names []string
-	err = chg.Get("snap-names", &names)
-	c.Assert(err, check.IsNil)
-	c.Check(names, check.DeepEquals, []string{"foo"})
-
-	c.Check(soon, check.Equals, 1)
-}
-
-func (s *apiSuite) TestPostSnapChannel(c *check.C) {
-	d := s.daemonWithOverlordMock(c)
-
-	soon := 0
-	ensureStateSoon = func(st *state.State) {
-		soon++
-		ensureStateSoonImpl(st)
-	}
-
-	s.vars = map[string]string{"name": "foo"}
-
-	snapInstructionDispTable["install"] = func(*snapInstruction, *state.State) (string, []*state.TaskSet, error) {
-		return "foooo", nil, nil
-	}
-	defer func() {
-		snapInstructionDispTable["install"] = snapInstall
-	}()
-
-	buf := bytes.NewBufferString(`{"action": "install"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-	c.Check(chg.Summary(), check.Equals, "foooo")
-	var names []string
-	err = chg.Get("snap-names", &names)
-	c.Assert(err, check.IsNil)
-	c.Check(names, check.DeepEquals, []string{"foo"})
-
-	c.Check(soon, check.Equals, 1)
-}
-
-func (s *apiSuite) TestPostSnapVerifySnapInstruction(c *check.C) {
-	s.daemonWithOverlordMock(c)
-
-	buf := bytes.NewBufferString(`{"action": "install"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/ubuntu-core", buf)
-	c.Assert(err, check.IsNil)
-	s.vars = map[string]string{"name": "ubuntu-core"}
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, `cannot install "ubuntu-core", please use "core" instead`)
-}
-
-func (s *apiSuite) TestPostSnapCohortRandoAction(c *check.C) {
-	s.daemonWithOverlordMock(c)
-	s.vars = map[string]string{"name": "some-snap"}
-	const expectedErr = "cohort-key can only be specified for install, refresh, or switch"
-
-	for _, action := range []string{"remove", "revert", "enable", "disable", "xyzzy"} {
-		buf := strings.NewReader(fmt.Sprintf(`{"action": "%s", "cohort-key": "32"}`, action))
-		req, err := http.NewRequest("POST", "/v2/snaps/some-snap", buf)
-		c.Assert(err, check.IsNil)
-
-		rsp := postSnap(snapCmd, req, nil).(*resp)
-
-		c.Check(rsp.Type, check.Equals, ResponseTypeError)
-		c.Check(rsp.Status, check.Equals, 400, check.Commentf("%q", action))
-		c.Check(rsp.Result.(*errorResult).Message, check.Equals, expectedErr, check.Commentf("%q", action))
-	}
-}
-
-func (s *apiSuite) TestPostSnapLeaveCohortRandoAction(c *check.C) {
-	s.daemonWithOverlordMock(c)
-	s.vars = map[string]string{"name": "some-snap"}
-	const expectedErr = "leave-cohort can only be specified for refresh or switch"
-
-	for _, action := range []string{"install", "remove", "revert", "enable", "disable", "xyzzy"} {
-		buf := strings.NewReader(fmt.Sprintf(`{"action": "%s", "leave-cohort": true}`, action))
-		req, err := http.NewRequest("POST", "/v2/snaps/some-snap", buf)
-		c.Assert(err, check.IsNil)
-
-		rsp := postSnap(snapCmd, req, nil).(*resp)
-
-		c.Check(rsp.Type, check.Equals, ResponseTypeError)
-		c.Check(rsp.Status, check.Equals, 400, check.Commentf("%q", action))
-		c.Check(rsp.Result.(*errorResult).Message, check.Equals, expectedErr, check.Commentf("%q", action))
-	}
-}
-
-func (s *apiSuite) TestPostSnapCohortIncompat(c *check.C) {
-	s.daemonWithOverlordMock(c)
-	s.vars = map[string]string{"name": "some-snap"}
-
-	type T struct {
-		opts   string
-		errmsg string
-	}
-
-	for i, t := range []T{
-		// TODO: more?
-		{`"cohort-key": "what", "revision": "42"`, `cannot specify both cohort-key and revision`},
-		{`"cohort-key": "what", "leave-cohort": true`, `cannot specify both cohort-key and leave-cohort`},
-	} {
-		buf := strings.NewReader(fmt.Sprintf(`{"action": "refresh", %s}`, t.opts))
-		req, err := http.NewRequest("POST", "/v2/snaps/some-snap", buf)
-		c.Assert(err, check.IsNil, check.Commentf("%d (%s)", i, t.opts))
-
-		rsp := postSnap(snapCmd, req, nil).(*resp)
-
-		c.Check(rsp.Type, check.Equals, ResponseTypeError, check.Commentf("%d (%s)", i, t.opts))
-		c.Check(rsp.Status, check.Equals, 400, check.Commentf("%d (%s)", i, t.opts))
-		c.Check(rsp.Result.(*errorResult).Message, check.Equals, t.errmsg, check.Commentf("%d (%s)", i, t.opts))
-	}
-}
-
-func (s *apiSuite) TestPostSnapSetsUser(c *check.C) {
-	d := s.daemon(c)
-	ensureStateSoon = func(st *state.State) {}
-
-	snapInstructionDispTable["install"] = func(inst *snapInstruction, st *state.State) (string, []*state.TaskSet, error) {
-		return fmt.Sprintf("<install by user %d>", inst.userID), nil, nil
-	}
-	defer func() {
-		snapInstructionDispTable["install"] = snapInstall
-	}()
-
-	state := snapCmd.d.overlord.State()
-	state.Lock()
-	user, err := auth.NewUser(state, "username", "email@test.com", "macaroon", []string{"discharge"})
-	state.Unlock()
-	c.Check(err, check.IsNil)
-
-	buf := bytes.NewBufferString(`{"action": "install"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Authorization", `Macaroon root="macaroon", discharge="discharge"`)
-
-	rsp := postSnap(snapCmd, req, user).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-	c.Check(chg.Summary(), check.Equals, "<install by user 1>")
-}
-
-func (s *apiSuite) TestPostSnapDispatch(c *check.C) {
-	inst := &snapInstruction{Snaps: []string{"foo"}}
-
-	type T struct {
-		s    string
-		impl snapActionFunc
-	}
-
-	actions := []T{
-		{"install", snapInstall},
-		{"refresh", snapUpdate},
-		{"remove", snapRemove},
-		{"revert", snapRevert},
-		{"enable", snapEnable},
-		{"disable", snapDisable},
-		{"switch", snapSwitch},
-		{"xyzzy", nil},
-	}
-
-	for _, action := range actions {
-		inst.Action = action.s
-		// do you feel dirty yet?
-		c.Check(fmt.Sprintf("%p", action.impl), check.Equals, fmt.Sprintf("%p", inst.dispatch()))
-	}
-}
-
-func (s *apiSuite) TestPostSnapEnableDisableSwitchRevision(c *check.C) {
-	for _, action := range []string{"enable", "disable", "switch"} {
-		buf := bytes.NewBufferString(`{"action": "` + action + `", "revision": "42"}`)
-		req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-		c.Assert(err, check.IsNil)
-
-		rsp := postSnap(snapCmd, req, nil).(*resp)
-
-		c.Check(rsp.Type, check.Equals, ResponseTypeError)
-		c.Check(rsp.Status, check.Equals, 400)
-		c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, "takes no revision")
-	}
-}
-
-func (s *apiSuite) TestInstallOnNonDevModeDistro(c *check.C) {
-	s.testInstall(c, false, snapstate.Flags{}, snap.R(0))
-}
-func (s *apiSuite) TestInstallOnDevModeDistro(c *check.C) {
-	s.testInstall(c, true, snapstate.Flags{}, snap.R(0))
-}
-func (s *apiSuite) TestInstallRevision(c *check.C) {
-	s.testInstall(c, false, snapstate.Flags{}, snap.R(42))
-}
-
-func (s *apiSuite) testInstall(c *check.C, forcedDevmode bool, flags snapstate.Flags, revision snap.Revision) {
-	calledFlags := snapstate.Flags{}
-	installQueue := []string{}
-	restore := sandbox.MockForceDevMode(forcedDevmode)
-	defer restore()
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-		installQueue = append(installQueue, name)
-		c.Check(revision, check.Equals, opts.Revision)
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	defer func() {
-		snapstateInstall = nil
-	}()
-
-	d := s.daemonWithFakeSnapManager(c)
-
-	var buf bytes.Buffer
-	if revision.Unset() {
-		buf.WriteString(`{"action": "install"}`)
-	} else {
-		fmt.Fprintf(&buf, `{"action": "install", "revision": %s}`, revision.String())
-	}
-	req, err := http.NewRequest("POST", "/v2/snaps/some-snap", &buf)
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "some-snap"}
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Assert(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-
-	c.Check(chg.Tasks(), check.HasLen, 1)
-
-	st.Unlock()
-	s.waitTrivialChange(c, chg)
-	st.Lock()
-
-	c.Check(chg.Status(), check.Equals, state.DoneStatus)
-	c.Check(calledFlags, check.Equals, flags)
-	c.Check(err, check.IsNil)
-	c.Check(installQueue, check.DeepEquals, []string{"some-snap"})
-	c.Check(chg.Kind(), check.Equals, "install-snap")
-	c.Check(chg.Summary(), check.Equals, `Install "some-snap" snap`)
-}
-
-func (s *apiSuite) TestInstallUserAgentContextCreated(c *check.C) {
-	snapstateInstall = func(ctx context.Context, st *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		s.ctx = ctx
-		t := st.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	defer func() {
-		snapstateInstall = nil
-	}()
-
-	s.daemonWithFakeSnapManager(c)
-
-	var buf bytes.Buffer
-	buf.WriteString(`{"action": "install"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/some-snap", &buf)
-	req.RemoteAddr = "pid=100;uid=0;socket=;"
-	c.Assert(err, check.IsNil)
-	req.Header.Add("User-Agent", "some-agent/1.0")
-
-	s.vars = map[string]string{"name": "some-snap"}
-	rec := httptest.NewRecorder()
-	snapCmd.ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	c.Check(store.ClientUserAgent(s.ctx), check.Equals, "some-agent/1.0")
-}
-
-func (s *apiSuite) TestRefresh(c *check.C) {
-	var calledFlags snapstate.Flags
-	calledUserID := 0
-	installQueue := []string{}
-	assertstateCalledUserID := 0
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-		calledUserID = userID
-		installQueue = append(installQueue, name)
-
-		t := s.NewTask("fake-refresh-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		assertstateCalledUserID = userID
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "refresh",
-		Snaps:  []string{"some-snap"},
-		userID: 17,
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(assertstateCalledUserID, check.Equals, 17)
-	c.Check(calledFlags, check.DeepEquals, snapstate.Flags{})
-	c.Check(calledUserID, check.Equals, 17)
-	c.Check(err, check.IsNil)
-	c.Check(installQueue, check.DeepEquals, []string{"some-snap"})
-	c.Check(summary, check.Equals, `Refresh "some-snap" snap`)
-}
-
-func (s *apiSuite) TestRefreshDevMode(c *check.C) {
-	var calledFlags snapstate.Flags
-	calledUserID := 0
-	installQueue := []string{}
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-		calledUserID = userID
-		installQueue = append(installQueue, name)
-
-		t := s.NewTask("fake-refresh-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:  "refresh",
-		DevMode: true,
-		Snaps:   []string{"some-snap"},
-		userID:  17,
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	flags := snapstate.Flags{}
-	flags.DevMode = true
-	c.Check(calledFlags, check.DeepEquals, flags)
-	c.Check(calledUserID, check.Equals, 17)
-	c.Check(err, check.IsNil)
-	c.Check(installQueue, check.DeepEquals, []string{"some-snap"})
-	c.Check(summary, check.Equals, `Refresh "some-snap" snap`)
-}
-
-func (s *apiSuite) TestRefreshClassic(c *check.C) {
-	var calledFlags snapstate.Flags
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-		return nil, nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:  "refresh",
-		Classic: true,
-		Snaps:   []string{"some-snap"},
-		userID:  17,
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(calledFlags, check.DeepEquals, snapstate.Flags{Classic: true})
-}
-
-func (s *apiSuite) TestRefreshIgnoreValidation(c *check.C) {
-	var calledFlags snapstate.Flags
-	calledUserID := 0
-	installQueue := []string{}
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-		calledUserID = userID
-		installQueue = append(installQueue, name)
-
-		t := s.NewTask("fake-refresh-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:           "refresh",
-		IgnoreValidation: true,
-		Snaps:            []string{"some-snap"},
-		userID:           17,
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	flags := snapstate.Flags{}
-	flags.IgnoreValidation = true
-
-	c.Check(calledFlags, check.DeepEquals, flags)
-	c.Check(calledUserID, check.Equals, 17)
-	c.Check(err, check.IsNil)
-	c.Check(installQueue, check.DeepEquals, []string{"some-snap"})
-	c.Check(summary, check.Equals, `Refresh "some-snap" snap`)
-}
-
-func (s *apiSuite) TestRefreshIgnoreRunning(c *check.C) {
-	var calledFlags snapstate.Flags
-	installQueue := []string{}
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-		installQueue = append(installQueue, name)
-
-		t := s.NewTask("fake-refresh-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:        "refresh",
-		IgnoreRunning: true,
-		Snaps:         []string{"some-snap"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	flags := snapstate.Flags{}
-	flags.IgnoreRunning = true
-
-	c.Check(calledFlags, check.DeepEquals, flags)
-	c.Check(err, check.IsNil)
-	c.Check(installQueue, check.DeepEquals, []string{"some-snap"})
-	c.Check(summary, check.Equals, `Refresh "some-snap" snap`)
-}
-
-func (s *apiSuite) TestRefreshCohort(c *check.C) {
-	cohort := ""
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		cohort = opts.CohortKey
-
-		t := s.NewTask("fake-refresh-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "refresh",
-		Snaps:  []string{"some-snap"},
-		snapRevisionOptions: snapRevisionOptions{
-			CohortKey: "xyzzy",
-		},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(cohort, check.Equals, "xyzzy")
-	c.Check(summary, check.Equals, `Refresh "some-snap" snap`)
-}
-
-func (s *apiSuite) TestRefreshLeaveCohort(c *check.C) {
-	var leave *bool
-
-	snapstateUpdate = func(s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		leave = &opts.LeaveCohort
-
-		t := s.NewTask("fake-refresh-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		return nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:              "refresh",
-		snapRevisionOptions: snapRevisionOptions{LeaveCohort: true},
-		Snaps:               []string{"some-snap"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(*leave, check.Equals, true)
-	c.Check(summary, check.Equals, `Refresh "some-snap" snap`)
-}
-
-func (s *apiSuite) TestSwitchInstruction(c *check.C) {
-	var cohort, channel string
-	var leave *bool
-	snapstateSwitch = func(s *state.State, name string, opts *snapstate.RevisionOptions) (*state.TaskSet, error) {
-		cohort = opts.CohortKey
-		leave = &opts.LeaveCohort
-		channel = opts.Channel
-
-		t := s.NewTask("fake-switch", "Doing a fake switch")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	st := d.overlord.State()
-
-	type T struct {
-		channel string
-		cohort  string
-		leave   bool
-		summary string
-	}
-	table := []T{
-		{"", "some-cohort", false, `Switch "some-snap" snap to cohort "…me-cohort"`},
-		{"some-channel", "", false, `Switch "some-snap" snap to channel "some-channel"`},
-		{"some-channel", "some-cohort", false, `Switch "some-snap" snap to channel "some-channel" and cohort "…me-cohort"`},
-		{"", "", true, `Switch "some-snap" snap away from cohort`},
-		{"some-channel", "", true, `Switch "some-snap" snap to channel "some-channel" and away from cohort`},
-	}
-
-	for _, t := range table {
-		cohort, channel = "", ""
-		leave = nil
-		inst := &snapInstruction{
-			Action: "switch",
-			snapRevisionOptions: snapRevisionOptions{
-				CohortKey:   t.cohort,
-				LeaveCohort: t.leave,
-				Channel:     t.channel,
-			},
-			Snaps: []string{"some-snap"},
-		}
-
-		st.Lock()
-		summary, _, err := inst.dispatch()(inst, st)
-		st.Unlock()
-		c.Check(err, check.IsNil)
-
-		c.Check(cohort, check.Equals, t.cohort)
-		c.Check(channel, check.Equals, t.channel)
-		c.Check(summary, check.Equals, t.summary)
-		c.Check(*leave, check.Equals, t.leave)
-	}
-}
-
-func (s *apiSuite) TestInstallFails(c *check.C) {
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		t := s.NewTask("fake-install-snap-error", "Install task")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemonWithFakeSnapManager(c)
-	s.vars = map[string]string{"name": "hello-world"}
-	buf := bytes.NewBufferString(`{"action": "install"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
-	c.Assert(err, check.IsNil)
-
-	rsp := postSnap(snapCmd, req, nil).(*resp)
-
-	c.Assert(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-
-	c.Check(chg.Tasks(), check.HasLen, 1)
-
-	st.Unlock()
-	s.waitTrivialChange(c, chg)
-	st.Lock()
-
-	c.Check(chg.Err(), check.ErrorMatches, `(?sm).*Install task \(fake-install-snap-error errored\)`)
-}
-
 func (s *apiSuite) TestInstallLeaveOld(c *check.C) {
 	c.Skip("temporarily dropped half-baked support while sorting out flag mess")
 	var calledFlags snapstate.Flags
@@ -1184,233 +471,6 @@ func (s *apiSuite) TestInstallLeaveOld(c *check.C) {
 
 	c.Check(calledFlags, check.DeepEquals, snapstate.Flags{})
 	c.Check(err, check.IsNil)
-}
-
-func (s *apiSuite) TestInstall(c *check.C) {
-	var calledName string
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledName = name
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "install",
-		// Install the snap in developer mode
-		DevMode: true,
-		Snaps:   []string{"fake"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-	c.Check(calledName, check.Equals, "fake")
-}
-
-func (s *apiSuite) TestInstallCohort(c *check.C) {
-	var calledName string
-	var calledCohort string
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledName = name
-		calledCohort = opts.CohortKey
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "install",
-		snapRevisionOptions: snapRevisionOptions{
-			CohortKey: "To the legion of the lost ones, to the cohort of the damned.",
-		},
-		Snaps: []string{"fake"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	msg, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-	c.Check(calledName, check.Equals, "fake")
-	c.Check(calledCohort, check.Equals, "To the legion of the lost ones, to the cohort of the damned.")
-	c.Check(msg, check.Equals, `Install "fake" snap from "…e damned." cohort`)
-}
-
-func (s *apiSuite) TestInstallDevMode(c *check.C) {
-	var calledFlags snapstate.Flags
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "install",
-		// Install the snap in developer mode
-		DevMode: true,
-		Snaps:   []string{"fake"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(calledFlags.DevMode, check.Equals, true)
-}
-
-func (s *apiSuite) TestInstallJailMode(c *check.C) {
-	var calledFlags snapstate.Flags
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:   "install",
-		JailMode: true,
-		Snaps:    []string{"fake"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(calledFlags.JailMode, check.Equals, true)
-}
-
-func (s *apiSuite) TestInstallJailModeDevModeOS(c *check.C) {
-	restore := sandbox.MockForceDevMode(true)
-	defer restore()
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:   "install",
-		JailMode: true,
-		Snaps:    []string{"foo"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.ErrorMatches, "this system cannot honour the jailmode flag")
-}
-
-func (s *apiSuite) TestInstallEmptyName(c *check.C) {
-	snapstateInstall = func(ctx context.Context, _ *state.State, _ string, _ *snapstate.RevisionOptions, _ int, _ snapstate.Flags) (*state.TaskSet, error) {
-		return nil, errors.New("should not be called")
-	}
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "install",
-		Snaps:  []string{""},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.ErrorMatches, "cannot install snap with empty name")
-}
-
-func (s *apiSuite) TestInstallJailModeDevMode(c *check.C) {
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action:   "install",
-		DevMode:  true,
-		JailMode: true,
-		Snaps:    []string{"foo"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.ErrorMatches, "cannot use devmode and jailmode flags together")
-}
-
-func (s *apiSuite) testRevertSnap(inst *snapInstruction, c *check.C) {
-	queue := []string{}
-
-	instFlags, err := inst.modeFlags()
-	c.Assert(err, check.IsNil)
-
-	snapstateRevert = func(s *state.State, name string, flags snapstate.Flags) (*state.TaskSet, error) {
-		c.Check(flags, check.Equals, instFlags)
-		queue = append(queue, name)
-		return nil, nil
-	}
-	snapstateRevertToRevision = func(s *state.State, name string, rev snap.Revision, flags snapstate.Flags) (*state.TaskSet, error) {
-		c.Check(flags, check.Equals, instFlags)
-		queue = append(queue, fmt.Sprintf("%s (%s)", name, rev))
-		return nil, nil
-	}
-
-	d := s.daemon(c)
-	inst.Action = "revert"
-	inst.Snaps = []string{"some-snap"}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	summary, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-	if inst.Revision.Unset() {
-		c.Check(queue, check.DeepEquals, []string{inst.Snaps[0]})
-	} else {
-		c.Check(queue, check.DeepEquals, []string{fmt.Sprintf("%s (%s)", inst.Snaps[0], inst.Revision)})
-	}
-	c.Check(summary, check.Equals, `Revert "some-snap" snap`)
-}
-
-func (s *apiSuite) TestRevertSnap(c *check.C) {
-	s.testRevertSnap(&snapInstruction{}, c)
-}
-
-func (s *apiSuite) TestRevertSnapDevMode(c *check.C) {
-	s.testRevertSnap(&snapInstruction{DevMode: true}, c)
-}
-
-func (s *apiSuite) TestRevertSnapJailMode(c *check.C) {
-	s.testRevertSnap(&snapInstruction{JailMode: true}, c)
-}
-
-func (s *apiSuite) TestRevertSnapClassic(c *check.C) {
-	s.testRevertSnap(&snapInstruction{Classic: true}, c)
-}
-
-func (s *apiSuite) TestRevertSnapToRevision(c *check.C) {
-	s.testRevertSnap(&snapInstruction{snapRevisionOptions: snapRevisionOptions{Revision: snap.R(1)}}, c)
-}
-
-func (s *apiSuite) TestRevertSnapToRevisionDevMode(c *check.C) {
-	s.testRevertSnap(&snapInstruction{snapRevisionOptions: snapRevisionOptions{Revision: snap.R(1)}, DevMode: true}, c)
-}
-
-func (s *apiSuite) TestRevertSnapToRevisionJailMode(c *check.C) {
-	s.testRevertSnap(&snapInstruction{snapRevisionOptions: snapRevisionOptions{Revision: snap.R(1)}, JailMode: true}, c)
-}
-
-func (s *apiSuite) TestRevertSnapToRevisionClassic(c *check.C) {
-	s.testRevertSnap(&snapInstruction{snapRevisionOptions: snapRevisionOptions{Revision: snap.R(1)}, Classic: true}, c)
 }
 
 func (s *apiSuite) TestIsTrue(c *check.C) {
