+++ daemon/export_test.go
@@ -125,11 +125,43 @@ func MockSnapstateInstallPath(mock func(*state.State, *snap.SideInfo, string, st
 	}
 }
 
+func MockSnapstateUpdate(mock func(*state.State, string, *snapstate.RevisionOptions, int, snapstate.Flags) (*state.TaskSet, error)) (restore func()) {
+	oldSnapstateUpdate := snapstateUpdate
+	snapstateUpdate = mock
+	return func() {
+		snapstateUpdate = oldSnapstateUpdate
+	}
+}
+
 func MockSnapstateTryPath(mock func(*state.State, string, string, snapstate.Flags) (*state.TaskSet, error)) (restore func()) {
 	oldSnapstateTryPath := snapstateTryPath
 	snapstateTryPath = mock
 	return func() {
 		snapstateTryPath = oldSnapstateTryPath
+	}
+}
+
+func MockSnapstateSwitch(mock func(*state.State, string, *snapstate.RevisionOptions) (*state.TaskSet, error)) (restore func()) {
+	oldSnapstateSwitch := snapstateSwitch
+	snapstateSwitch = mock
+	return func() {
+		snapstateSwitch = oldSnapstateSwitch
+	}
+}
+
+func MockSnapstateRevert(mock func(*state.State, string, snapstate.Flags) (*state.TaskSet, error)) (restore func()) {
+	oldSnapstateRevert := snapstateRevert
+	snapstateRevert = mock
+	return func() {
+		snapstateRevert = oldSnapstateRevert
+	}
+}
+
+func MockSnapstateRevertToRevision(mock func(*state.State, string, snap.Revision, snapstate.Flags) (*state.TaskSet, error)) (restore func()) {
+	oldSnapstateRevertToRevision := snapstateRevertToRevision
+	snapstateRevertToRevision = mock
+	return func() {
+		snapstateRevertToRevision = oldSnapstateRevertToRevision
 	}
 }
 
@@ -169,4 +201,12 @@ func (inst *snapInstruction) Dispatch() snapActionFunc {
 
 func (inst *snapInstruction) DispatchForMany() snapManyActionFunc {
 	return inst.dispatchForMany()
+}
+
+func (inst *snapInstruction) SetUserID(userID int) {
+	inst.userID = userID
+}
+
+func (inst *snapInstruction) ModeFlags() (snapstate.Flags, error) {
+	return inst.modeFlags()
 }
